pyproject.toml
[tool.poetry]
name = "platform-vector"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.40"
psycopg2-binary = "^2.9.10"
pydantic = "^2.11.2"
python-dotenv = "^1.1.0"
openai = "^1.70.0"
pydantic-settings = "^2.8.1"
numpy = "^2.2.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
